requirements.txt
fastapi==0.78.0
uvicorn==0.17.6
requests==2.27.1
passlib==1.7.4
python-multipart==0.0.5
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
oauth2client